+++ GUI/src/lib/utils.js
@@ -56,3 +56,38 @@ export let wait = (time) => {
         setTimeout(() => resolve(), time);
     });
 }
+
+export let pathsFromTree = (tree) => {
+    let flatpaths = [];
+    let path = [];
+    let pathFromTree = (tree) => {
+        for (const key of Object.keys(tree)) {
+            path.push(key);
+            if (tree[key] != null && typeof tree[key] == "object") {
+              pathFromTree(tree[key]);
+            }
+            else {
+              flatpaths.push(path.join('.'));
+            }
+            path.pop();
+        }
+    }
+    // array path
+    pathFromTree(tree);
+    return flatpaths;
+}
+
+export let numberDisplay = (val) => {
+    // if a number can be represented with 3 decimals, return it in that form
+    // otherwise, return scientific notation
+    let retVal = '';
+    try {
+        retVal = parseFloat(val).toFixed(3);
+        if (retVal == '0.000' && val != 0 || retVal.length > 7) {
+            retVal = parseFloat(val).toExponential(3);
+        }
+    } catch (error) {
+        console.log(error);
+    }
+    return retVal;
+}